+++ src/lightning/pytorch/core/datamodule.py
@@ -85,8 +85,7 @@ class LightningDataModule(DataHooks, HyperparametersMixin):
         num_workers: int = 0,
         **datamodule_kwargs: Any,
     ) -> "LightningDataModule":
-        r"""
-        Create an instance from torch.utils.data.Dataset.
+        r"""Create an instance from torch.utils.data.Dataset.
 
         Args:
             train_dataset: Optional dataset or iterable of datasets to be used for train_dataloader()