+++ src/lightning/pytorch/loggers/neptune.py
@@ -61,8 +61,7 @@ _INTEGRATION_VERSION_KEY = "source_code/integrations/pytorch-lightning"
 
 
 class NeptuneLogger(Logger):
-    r"""
-    Log using `Neptune <https://neptune.ai>`_.
+    r"""Log using `Neptune <https://neptune.ai>`_.
 
     Install it with pip:
 
@@ -295,11 +294,9 @@ class NeptuneLogger(Logger):
             args["run"] = self._run_short_id
 
         # Backward compatibility in case of previous version retrieval
-        try:
+        with contextlib.suppress(AttributeError):
             if self._run_name is not None:
                 args["name"] = self._run_name
-        except AttributeError:
-            pass
 
         return args
 
@@ -382,8 +379,7 @@ class NeptuneLogger(Logger):
 
     @rank_zero_only
     def log_hyperparams(self, params: Union[Dict[str, Any], Namespace]) -> None:  # skipcq: PYL-W0221
-        r"""
-        Log hyperparameters to the run.
+        r"""Log hyperparameters to the run.
 
         Hyperparameters will be logged under the "<prefix>/hyperparams" namespace.