+++ tests/tests_pytorch/accelerators/test_tpu.py
@@ -100,7 +100,6 @@ def test_accelerator_tpu(accelerator, devices, tpu_available):
 
 
 class ManualOptimizationModel(BoringModel):
-
     count = 0
     called = collections.defaultdict(int)
 
@@ -193,7 +192,6 @@ def test_strategy_choice_tpu_strategy():
 @RunIf(tpu=True)
 @mock.patch.dict(os.environ, os.environ.copy(), clear=True)
 def test_auto_parameters_tying_tpus(tmpdir):
-
     model = WeightSharingModule()
     shared_params = find_shared_parameters(model)